refactor(docs): restructure badge plugin architecture with modular design
