+++ docs/.vitepress/plugins/badge/constants.ts
@@ -1,0 +1,43 @@
+/**
+ * Badge 相关常量配置
+ */
+
+/**
+ * Badge 类型定义
+ */
+export type BadgeType = 'new' | 'deprecated' | 'beta' | 'alpha'
+
+/**
+ * Badge 值类型（包含固定类型和版本号字符串）
+ */
+export type BadgeValue = BadgeType | string
+
+/**
+ * Badge 类型到中文文本的映射
+ */
+export const BADGE_TEXT_MAP: Record<BadgeType, string> = {
+  new: '新增',
+  deprecated: '已废弃',
+  beta: 'Beta',
+  alpha: 'Alpha',
+}
+
+/**
+ * Badge CSS 类名映射
+ */
+export const BADGE_CLASS_MAP: Record<BadgeType, string> = {
+  new: 'version-badge version-badge--new',
+  deprecated: 'version-badge version-badge--deprecated',
+  beta: 'version-badge version-badge--beta',
+  alpha: 'version-badge version-badge--alpha',
+}
+
+/**
+ * 版本号正则表达式（支持 v0.4.0 或 0.4.0 格式）
+ */
+export const VERSION_NUMBER_REGEX = /^v?[\d.]+(?:[-+][a-zA-Z0-9.]+)?$/
+
+/**
+ * Markdown 中的版本标记正则表达式（用于 @new、@1.2.0 等）
+ */
+export const MARKDOWN_BADGE_REGEX = /@(new|deprecated|beta|alpha|[\d.]+(?:[-+][a-zA-Z0-9.]+)?)/g
