+++ docs/.vitepress/plugins/badge/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Badge 插件统一导出
+ */
+
+export { SidebarBadgePlugin } from './SidebarPlugin'
+export { MarkdownBadgePlugin } from './MarkdownPlugin'
+export * from './constants'
+export * from './utils'
+
+export type { SidebarBadgeOptions } from './SidebarPlugin'
