+++ docs/package.json
@@ -7,6 +7,7 @@
     "preview": "vitepress preview"
   },
   "devDependencies": {
+    "@types/node": "^22.10.2",
     "@vitejs/plugin-vue-jsx": "^4.1.2",
     "cross-env": "^7.0.3",
     "gh-pages": "^2.1.1",
